chore!(code): Remove VoteSet-based synchronization protocol
