+++ code/crates/core-consensus/src/handle.rs
@@ -11,7 +11,6 @@ mod sync;
 mod timeout;
 mod validator_set;
 mod vote;
-mod vote_set;
 
 use proposal::on_proposal;
 use propose::on_propose;
@@ -20,7 +19,6 @@ use start_height::reset_and_start_height;
 use sync::on_commit_certificate;
 use timeout::on_timeout_elapsed;
 use vote::on_vote;
-use vote_set::{on_vote_set_request, on_vote_set_response};
 
 use crate::prelude::*;
 
@@ -60,12 +58,6 @@ where
         }
         Input::CommitCertificate(certificate) => {
             on_commit_certificate(co, state, metrics, certificate).await
-        }
-        Input::VoteSetRequest(request_id, height, round) => {
-            on_vote_set_request(co, state, metrics, request_id, height, round).await
-        }
-        Input::VoteSetResponse(vote_set, polka_certificate) => {
-            on_vote_set_response(co, state, metrics, vote_set, polka_certificate).await
         }
     }
 }
