+++ code/crates/core-consensus/src/input.rs
@@ -1,13 +1,9 @@
 use derive_where::derive_where;
 use malachitebft_core_types::{
-    CommitCertificate, Context, PolkaCertificate, Round, SignedProposal, SignedVote, Timeout,
-    ValueOrigin, VoteSet,
+    CommitCertificate, Context, SignedProposal, SignedVote, Timeout, ValueOrigin,
 };
 
-use crate::types::ProposedValue;
-use crate::LocallyProposedValue;
-
-pub type RequestId = String;
+use crate::types::{LocallyProposedValue, ProposedValue};
 
 /// Inputs to be handled by the consensus process.
 #[derive_where(Clone, Debug, PartialEq, Eq)]
@@ -42,10 +38,4 @@ where
 
     /// We have received a commit certificate via the sync protocol.
     CommitCertificate(CommitCertificate<Ctx>),
-
-    /// A peer has requested a vote set for the given height and round
-    VoteSetRequest(RequestId, Ctx::Height, Round),
-
-    /// Received a vote set from a peer via Sync
-    VoteSetResponse(VoteSet<Ctx>, Vec<PolkaCertificate<Ctx>>),
 }
