+++ code/crates/engine/src/consensus.rs
@@ -14,16 +14,14 @@ use tracing::{debug, error, error_span, info, warn};
 use malachitebft_codec as codec;
 use malachitebft_config::TimeoutConfig;
 use malachitebft_core_consensus::{
-    Effect, PeerId, Resumable, Resume, SignedConsensusMsg, VoteExtensionError, VoteSyncMode,
+    Effect, PeerId, Resumable, Resume, SignedConsensusMsg, VoteExtensionError,
 };
 use malachitebft_core_types::{
     Context, Proposal, Round, SigningProvider, SigningProviderExt, Timeout, TimeoutKind,
     ValidatorSet, ValueId, ValueOrigin, Vote,
 };
 use malachitebft_metrics::Metrics;
-use malachitebft_sync::{
-    self as sync, InboundRequestId, Response, ValueResponse, VoteSetRequest, VoteSetResponse,
-};
+use malachitebft_sync::{self as sync, ValueResponse};
 
 use crate::host::{HostMsg, HostRef, LocallyProposedValue, ProposedValue};
 use crate::network::{NetworkEvent, NetworkMsg, NetworkRef};
@@ -503,58 +501,6 @@ where
                             },
                             None,
                         )?;
-                    }
-
-                    NetworkEvent::Request(
-                        request_id,
-                        peer,
-                        sync::Request::VoteSetRequest(VoteSetRequest { height, round }),
-                    ) => {
-                        debug!(%height, %round, %request_id, %peer, "Received vote set request");
-
-                        if let Err(e) = self
-                            .process_input(
-                                &myself,
-                                state,
-                                ConsensusInput::VoteSetRequest(
-                                    request_id.to_string(),
-                                    height,
-                                    round,
-                                ),
-                            )
-                            .await
-                        {
-                            error!(%peer, %height, %round, "Error when processing VoteSetRequest: {e:?}");
-                        }
-                    }
-
-                    NetworkEvent::Response(
-                        request_id,
-                        peer,
-                        sync::Response::VoteSetResponse(VoteSetResponse {
-                            height,
-                            round,
-                            vote_set,
-                            polka_certificates,
-                        }),
-                    ) => {
-                        if vote_set.votes.is_empty() {
-                            debug!(%height, %round, %request_id, %peer, "Received an empty vote set response");
-                            return Ok(());
-                        };
-
-                        debug!(%height, %round, %request_id, %peer, "Received a non-empty vote set response");
-
-                        if let Err(e) = self
-                            .process_input(
-                                &myself,
-                                state,
-                                ConsensusInput::VoteSetResponse(vote_set, polka_certificates),
-                            )
-                            .await
-                        {
-                            error!(%height, %round, %request_id, %peer, "Error when processing VoteSetResponse: {e:?}");
-                        }
                     }
 
                     NetworkEvent::Vote(from, vote) => {
@@ -1073,16 +1019,13 @@ where
             }
 
             Effect::Rebroadcast(msg, r) => {
-                // Rebroadcast last vote only if vote sync mode is set to "rebroadcast",
-                // otherwise vote set requests are issued automatically by the sync protocol.
-                if self.params.vote_sync_mode == VoteSyncMode::Rebroadcast {
-                    // Notify any subscribers that we are about to rebroadcast a message
-                    self.tx_event.send(|| Event::Rebroadcast(msg.clone()));
+                // Notify any subscribers that we are about to rebroadcast a message
+                self.tx_event.send(|| Event::Rebroadcast(msg.clone()));
 
-                    self.network
-                        .cast(NetworkMsg::Publish(SignedConsensusMsg::Vote(msg)))
-                        .map_err(|e| eyre!("Error when rebroadcasting vote message: {e:?}"))?;
-                }
+                // Rebroadcast our latest vote
+                self.network
+                    .cast(NetworkMsg::Publish(SignedConsensusMsg::Vote(msg)))
+                    .map_err(|e| eyre!("Error when rebroadcasting vote message: {e:?}"))?;
 
                 Ok(r.resume_with(()))
             }
@@ -1145,65 +1088,6 @@ where
                     sync.cast(SyncMsg::Decided(height))
                         .map_err(|e| eyre!("Error when sending decided height to sync: {e:?}"))?;
                 }
-
-                Ok(r.resume_with(()))
-            }
-
-            Effect::RequestVoteSet(height, round, r) => {
-                if let Some(sync) = &self.sync {
-                    debug!(%height, %round, "Request sync to obtain the vote set from peers");
-
-                    sync.cast(SyncMsg::RequestVoteSet(height, round))
-                        .map_err(|e| eyre!("Error when sending vote set request to sync: {e:?}"))?;
-
-                    self.tx_event
-                        .send(|| Event::RequestedVoteSet(height, round));
-                }
-
-                Ok(r.resume_with(()))
-            }
-
-            Effect::SendVoteSetResponse(
-                request_id_str,
-                height,
-                round,
-                vote_set,
-                polka_certificates,
-                r,
-            ) => {
-                let Some(sync) = self.sync.as_ref() else {
-                    warn!("Responding to a vote set request but sync actor is not available");
-                    return Ok(r.resume_with(()));
-                };
-
-                let vote_count = vote_set.len();
-                let polka_certificates_count = polka_certificates.len();
-
-                let response = Response::VoteSetResponse(VoteSetResponse::new(
-                    height,
-                    round,
-                    vote_set,
-                    polka_certificates,
-                ));
-
-                let request_id = InboundRequestId::new(request_id_str);
-
-                debug!(
-                    %height, %round, %request_id, vote.count = %vote_count,
-                    "Sending the vote set response"
-                );
-
-                self.network
-                    .cast(NetworkMsg::OutgoingResponse(request_id.clone(), response))?;
-
-                sync.cast(SyncMsg::SentVoteSetResponse(request_id, height, round))
-                    .map_err(|e| {
-                        eyre!("Error when notifying Sync about vote set response: {e:?}")
-                    })?;
-
-                self.tx_event.send(|| {
-                    Event::SentVoteSetResponse(height, round, vote_count, polka_certificates_count)
-                });
 
                 Ok(r.resume_with(()))
             }
