+++ code/crates/starknet/host/src/spawn.rs
@@ -4,10 +4,7 @@ use std::time::Duration;
 use tokio::task::JoinHandle;
 use tracing::warn;
 
-use malachitebft_config::{
-    self as config, MempoolConfig, MempoolLoadConfig, ValueSyncConfig, VoteSyncConfig,
-};
-use malachitebft_core_consensus::VoteSyncMode;
+use malachitebft_config::{self as config, MempoolConfig, MempoolLoadConfig, ValueSyncConfig};
 use malachitebft_core_types::ValuePayload;
 use malachitebft_engine::consensus::{Consensus, ConsensusParams, ConsensusRef};
 use malachitebft_engine::host::HostRef;
@@ -78,7 +75,6 @@ pub async fn spawn_node_actor(
         network.clone(),
         host.clone(),
         &cfg.value_sync,
-        &cfg.consensus.vote_sync,
         &registry,
         &span,
     )
@@ -133,11 +129,10 @@ async fn spawn_sync_actor(
     network: NetworkRef<MockContext>,
     host: HostRef<MockContext>,
     config: &ValueSyncConfig,
-    vote_sync: &VoteSyncConfig,
     registry: &SharedRegistry,
     span: &tracing::Span,
 ) -> Option<SyncRef<MockContext>> {
-    if !config.enabled && vote_sync.mode != config::VoteSyncMode::RequestResponse {
+    if !config.enabled {
         return None;
     }
 
@@ -176,10 +171,6 @@ async fn spawn_consensus_actor(
         address,
         threshold_params: Default::default(),
         value_payload: ValuePayload::PartsOnly,
-        vote_sync_mode: match cfg.consensus.vote_sync.mode {
-            config::VoteSyncMode::RequestResponse => VoteSyncMode::RequestResponse,
-            config::VoteSyncMode::Rebroadcast => VoteSyncMode::Rebroadcast,
-        },
     };
 
     Consensus::spawn(
