+++ code/crates/starknet/p2p-proto/proto/sync.proto
@@ -78,34 +78,14 @@ message ProposedValue {
     bool validity = 7;
 }
 
-message VoteSetRequest {
-  uint64 fork_id = 1;
-  uint64 block_number = 2;
-  uint32 round = 3;
-}
-
-message VoteSetResponse {
-  uint64 fork_id = 1;
-  uint64 block_number = 2;
-  uint32 round = 3;
-  VoteSet vote_set = 4;
-  repeated PolkaCertificate polka_certificates = 5;
-}
-
-message VoteSet {
-  repeated Vote signed_votes = 1;
-}
-
 message SyncRequest {
   oneof messages {
     ValueRequest value_request = 1;
-    VoteSetRequest vote_set_request = 2;
   }
 }
 
 message SyncResponse {
   oneof messages {
     ValueResponse value_response = 1;
-    VoteSetResponse vote_set_response = 2;
   }
 }
