+++ code/crates/starknet/test/src/lib.rs
@@ -146,9 +146,6 @@ impl TestRunner {
             logging: LoggingConfig::default(),
             consensus: ConsensusConfig {
                 value_payload: ValuePayload::PartsOnly,
-                vote_sync: VoteSyncConfig {
-                    mode: VoteSyncMode::Rebroadcast,
-                },
                 timeouts: TimeoutConfig::default(),
                 p2p: P2pConfig {
                     protocol,
@@ -239,8 +236,4 @@ fn apply_params(config: &mut Config, params: &TestParams) {
     config.test.vote_extensions.enabled = params.vote_extensions.is_some();
     config.test.vote_extensions.size = params.vote_extensions.unwrap_or_default();
     config.test.max_retain_blocks = params.max_retain_blocks;
-
-    if let Some(vote_sync_mode) = params.vote_sync_mode {
-        config.consensus.vote_sync.mode = vote_sync_mode;
-    }
 }
