+++ code/crates/starknet/test/src/tests/mod.rs
@@ -3,6 +3,5 @@ pub mod n3f0;
 pub mod n3f0_pubsub_protocol;
 pub mod n3f1;
 pub mod value_sync;
-pub mod vote_sync;
-pub mod vote_sync_bcast;
+pub mod vote_rebroadcast;
 pub mod wal;
