+++ code/crates/sync/src/types.rs
@@ -4,7 +4,7 @@ use displaydoc::Display;
 use libp2p::request_response;
 use serde::{Deserialize, Serialize};
 
-use malachitebft_core_types::{CommitCertificate, Context, PolkaCertificate, Round, VoteSet};
+use malachitebft_core_types::{CommitCertificate, Context};
 pub use malachitebft_peer::PeerId;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Display)]
@@ -39,13 +39,11 @@ pub struct Status<Ctx: Context> {
 #[derive_where(Clone, Debug, PartialEq, Eq)]
 pub enum Request<Ctx: Context> {
     ValueRequest(ValueRequest<Ctx>),
-    VoteSetRequest(VoteSetRequest<Ctx>),
 }
 
 #[derive_where(Clone, Debug, PartialEq, Eq)]
 pub enum Response<Ctx: Context> {
     ValueResponse(ValueResponse<Ctx>),
-    VoteSetResponse(VoteSetResponse<Ctx>),
 }
 
 #[derive_where(Clone, Debug, PartialEq, Eq)]
@@ -105,43 +103,3 @@ pub struct RawRequest(pub Bytes);
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct RawResponse(pub Bytes);
-
-#[derive_where(Clone, Debug, PartialEq, Eq)]
-pub struct VoteSetRequest<Ctx: Context> {
-    pub height: Ctx::Height,
-    pub round: Round,
-}
-
-impl<Ctx: Context> VoteSetRequest<Ctx> {
-    pub fn new(height: Ctx::Height, round: Round) -> Self {
-        Self { height, round }
-    }
-}
-
-#[derive_where(Clone, Debug, PartialEq, Eq)]
-pub struct VoteSetResponse<Ctx: Context> {
-    /// The height of the vote set
-    pub height: Ctx::Height,
-    /// The round of the vote set
-    pub round: Round,
-    /// The set of votes at this height and round
-    pub vote_set: VoteSet<Ctx>,
-    /// Certificates witnessing a Polka at this height for any round up to this round included
-    pub polka_certificates: Vec<PolkaCertificate<Ctx>>,
-}
-
-impl<Ctx: Context> VoteSetResponse<Ctx> {
-    pub fn new(
-        height: Ctx::Height,
-        round: Round,
-        vote_set: VoteSet<Ctx>,
-        polka_certificates: Vec<PolkaCertificate<Ctx>>,
-    ) -> Self {
-        Self {
-            height,
-            round,
-            vote_set,
-            polka_certificates,
-        }
-    }
-}
