+++ code/crates/test/app/src/node.rs
@@ -234,9 +234,6 @@ fn make_config(index: usize, total: usize, settings: MakeConfigSettings) -> Conf
         consensus: ConsensusConfig {
             // Current test app does not support proposal-only value payload properly as Init does not include valid_round
             value_payload: ValuePayload::ProposalAndParts,
-            vote_sync: VoteSyncConfig {
-                mode: VoteSyncMode::RequestResponse,
-            },
             timeouts: TimeoutConfig::default(),
             p2p: P2pConfig {
                 protocol: PubSubProtocol::default(),
