+++ code/crates/test/framework/src/params.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 use bytesize::ByteSize;
 
-use malachitebft_config::{PubSubProtocol, ValuePayload, VoteSyncMode};
+use malachitebft_config::{PubSubProtocol, ValuePayload};
 use malachitebft_test_app::config::Config;
 
 #[derive(Copy, Clone, Debug)]
@@ -16,7 +16,6 @@ pub struct TestParams {
     pub value_payload: ValuePayload,
     pub max_retain_blocks: usize,
     pub timeout_step: Duration,
-    pub vote_sync_mode: Option<VoteSyncMode>,
     pub stable_block_times: bool,
 }
 
@@ -32,7 +31,6 @@ impl Default for TestParams {
             value_payload: ValuePayload::ProposalAndParts,
             max_retain_blocks: 50,
             timeout_step: Duration::from_secs(2),
-            vote_sync_mode: None,
             stable_block_times: true,
         }
     }
@@ -50,9 +48,5 @@ impl TestParams {
         config.test.vote_extensions.size = self.vote_extensions.unwrap_or_default();
         config.test.max_retain_blocks = self.max_retain_blocks;
         config.test.stable_block_times = self.stable_block_times;
-
-        if let Some(vote_sync_mode) = self.vote_sync_mode {
-            config.consensus.vote_sync.mode = vote_sync_mode;
-        }
     }
 }
