+++ code/crates/test/tests/it/main.rs
@@ -6,8 +6,7 @@ mod n3f1;
 mod reset;
 mod validator_set;
 mod value_sync;
-mod vote_sync;
-mod vote_sync_bcast;
+mod vote_rebroadcast;
 mod wal;
 
 use std::collections::HashMap;
@@ -143,9 +142,6 @@ impl TestRunner {
             consensus: ConsensusConfig {
                 // Current test app does not support proposal-only value payload properly as Init does not include valid_round
                 value_payload: ValuePayload::ProposalAndParts,
-                vote_sync: VoteSyncConfig {
-                    mode: VoteSyncMode::RequestResponse,
-                },
                 timeouts: TimeoutConfig {
                     timeout_step: Duration::from_secs(2),
                     ..Default::default()
