+++ code/crates/test/tests/it/vote_rebroadcast.rs
@@ -1,33 +1,43 @@
 use std::time::Duration;
 
-use malachitebft_config::{ValuePayload, VoteSyncMode};
+use malachitebft_config::ValuePayload;
 
 use crate::{TestBuilder, TestParams};
 
 // NOTE: These tests are very similar to the Sync tests, with the difference that
 //       all nodes have the same voting power and therefore get stuck when one of them dies.
 
-pub async fn crash_restart_from_start(params: TestParams) {
-    const HEIGHT: u64 = 10;
+#[tokio::test]
+pub async fn crash_restart_from_start() {
     const CRASH_HEIGHT: u64 = 4;
+    const HEIGHT: u64 = 10;
 
     let mut test = TestBuilder::<()>::new();
 
-    test.add_node().start().wait_until(HEIGHT).success();
-    test.add_node().start().wait_until(HEIGHT).success();
+    test.add_node()
+        .start()
+        .wait_until(CRASH_HEIGHT)
+        .expect_vote_rebroadcast(CRASH_HEIGHT)
+        .wait_until(HEIGHT)
+        .success();
+
+    test.add_node()
+        .start()
+        .wait_until(CRASH_HEIGHT)
+        .expect_vote_rebroadcast(CRASH_HEIGHT)
+        .wait_until(HEIGHT)
+        .success();
 
     test.add_node()
         .start()
         // Wait until the node reaches height 4...
-        .wait_until(CRASH_HEIGHT)
+        .wait_until(4)
         // ...then kill it
         .crash()
         // Reset the database so that the node has to do Sync from height 1
         .reset_db()
         // After that, it waits 5 seconds before restarting the node
         .restart_after(Duration::from_secs(5))
-        // Expect a vote set request for height 4
-        .expect_vote_set_request(CRASH_HEIGHT)
         // Wait until the node reached the expected height
         .wait_until(HEIGHT)
         // Record a successful test for this node
@@ -35,44 +45,16 @@ pub async fn crash_restart_from_start(params: TestParams) {
 
     test.build()
         .run_with_params(
-            Duration::from_secs(60), // Timeout for the whole test
+            Duration::from_secs(60),
             TestParams {
-                vote_sync_mode: Some(VoteSyncMode::RequestResponse),
-                ..params
+                // Enable ValueSync to allow the node to catch up to the latest height
+                enable_value_sync: true,
+                timeout_step: Duration::from_secs(5),
+                value_payload: ValuePayload::PartsOnly,
+                ..TestParams::default()
             },
         )
         .await
-}
-
-#[tokio::test]
-pub async fn crash_restart_from_start_parts_only() {
-    let params = TestParams {
-        value_payload: ValuePayload::PartsOnly,
-        ..Default::default()
-    };
-
-    crash_restart_from_start(params).await
-}
-
-#[tokio::test]
-#[ignore] // Test app does not support proposal-only mode
-pub async fn crash_restart_from_start_proposal_only() {
-    let params = TestParams {
-        value_payload: ValuePayload::ProposalOnly,
-        ..Default::default()
-    };
-
-    crash_restart_from_start(params).await
-}
-
-#[tokio::test]
-pub async fn crash_restart_from_start_proposal_and_parts() {
-    let params = TestParams {
-        value_payload: ValuePayload::ProposalAndParts,
-        ..Default::default()
-    };
-
-    crash_restart_from_start(params).await
 }
 
 #[tokio::test]
@@ -85,12 +67,14 @@ pub async fn crash_restart_from_latest() {
     test.add_node()
         .start()
         .wait_until(CRASH_HEIGHT)
+        .expect_vote_rebroadcast(CRASH_HEIGHT)
         .wait_until(HEIGHT)
         .success();
 
     test.add_node()
         .start()
         .wait_until(CRASH_HEIGHT)
+        .expect_vote_rebroadcast(CRASH_HEIGHT)
         .wait_until(HEIGHT)
         .success();
 
@@ -100,7 +84,6 @@ pub async fn crash_restart_from_latest() {
         .crash()
         // We do not reset the database so that the node can restart from the latest height
         .restart_after(Duration::from_secs(5))
-        .expect_vote_set_request(CRASH_HEIGHT)
         .wait_until(HEIGHT)
         .success();
 
@@ -108,7 +91,7 @@ pub async fn crash_restart_from_latest() {
         .run_with_params(
             Duration::from_secs(60),
             TestParams {
-                vote_sync_mode: Some(VoteSyncMode::RequestResponse),
+                enable_value_sync: false,
                 ..Default::default()
             },
         )
@@ -121,12 +104,22 @@ pub async fn start_late() {
 
     let mut test = TestBuilder::<()>::new();
 
-    test.add_node().start().wait_until(HEIGHT).success();
-    test.add_node().start().wait_until(HEIGHT).success();
+    test.add_node()
+        .start()
+        .wait_until(1)
+        .expect_vote_rebroadcast(1)
+        .wait_until(HEIGHT)
+        .success();
+
+    test.add_node()
+        .start()
+        .wait_until(1)
+        .expect_vote_rebroadcast(1)
+        .wait_until(HEIGHT)
+        .success();
 
     test.add_node()
         .start_after(1, Duration::from_secs(10))
-        .expect_vote_set_request(1)
         .wait_until(HEIGHT)
         .success();
 
@@ -134,8 +127,8 @@ pub async fn start_late() {
         .run_with_params(
             Duration::from_secs(60),
             TestParams {
-                enable_value_sync: true, // Enable ValueSync to allow node to catch up to latest height
-                vote_sync_mode: Some(VoteSyncMode::RequestResponse),
+                // Enable ValueSync to allow the node to catch up to the latest height
+                enable_value_sync: true,
                 ..Default::default()
             },
         )
