+++ code/examples/channel/src/node.rs
@@ -227,9 +227,6 @@ fn make_config(index: usize, total: usize, settings: MakeConfigSettings) -> Conf
         consensus: ConsensusConfig {
             // Current channel app does not support parts-only value payload properly as Init does not include valid_round
             value_payload: ValuePayload::ProposalAndParts,
-            vote_sync: VoteSyncConfig {
-                mode: VoteSyncMode::RequestResponse,
-            },
             timeouts: TimeoutConfig::default(),
             p2p: P2pConfig {
                 protocol: PubSubProtocol::default(),
